[CAP-005] 키워드 컬렉션뷰 좌측정렬 구현
